Move `parse_page()` from `page.py` to `extractor/en/page.py`
and load `parse_page()` according to language code
